requirements.txt
pip>=18.0
lxml>=4.2.0
pytest
pillow
tox
pandas
xmlschema
